Вынес методы для работы с telegram bot api в отдельный класс
